Feature/merge fixes from arlm (#3)
* Error message showing the character instead of the character hexadecimal code

* Fixing PdfnameObject to handle null references on the equality operators

* DP operator should behave exactly as BDC operator. The standard states they expect similar parameters on the same pattern, eg.: /Name <<Property Dictionary>> DP

* Properly handling references pointing to null (as a value on the PDF)

* Properly handling null values on the Document Information dictionary

* Properly handling source code as text files, not binaries

* Properly handling Hex strings inside dictionaries

* Properly handling floats with too many floating-point digits

* Fix namespace and encoding

+++ src/PdfSharper/Internal/Diagnostics.cs
@@ -110,7 +110,7 @@ namespace PdfSharper.Internal
         {
             string message = string.Format(CultureInfo.InvariantCulture,
                 "Unexpected character '0x{0:x4}' in content stream. The stream may be corrupted or the feature is not implemented. " +
-                "If you think this is a bug in PDFsharp, please send us your PDF file.", ch);
+                "If you think this is a bug in PDFsharp, please send us your PDF file.", (int)ch);
             ThrowContentReaderException(message);
         }
     }
